add doc footer for edit guide, was this helpful, direction links
