+++ src/templates/doc-template.tsx
@@ -3,8 +3,8 @@ import { Helmet } from 'react-helmet';
 import Layout from '../components/Layout/Layout';
 import RightMenu from '../components/Layout/RightMenu';
 import '../styles/doc-template.css';
-import { Button } from '@material-ui/core';
-import { Link, graphql } from 'gatsby';
+import { graphql } from 'gatsby';
+import  DocFooter  from '../components/Layout/DocFooter';
 import { groupBy as _groupBy, forEach as _forEach, flatten as _flatten } from 'lodash';
 
 
@@ -14,20 +14,6 @@ const Template = ({ data }) => {
   const sectionsWithGuides = _groupBy(data.allMarkdownRemark.edges, 'node.frontmatter.section');
   let contentsArray = [];
   _forEach(sectionsWithGuides, (section, title) => contentsArray = [...contentsArray, {title: title, sections: section.map((s) => s.node)}]);
-
-  const gitHubUrl = 'https://github.com/ordercloud-api/oc-documentation/tree/development/src/pages/docs';
-  const flatContents = _flatten(contentsArray.map((c) => c.sections));
-  const guideIndex = flatContents.findIndex((section) => section.frontmatter.path === post.frontmatter.path);
-
-  const directionalButton = (direction) => {
-    const newGuideIndex = direction === 'Previous' ? guideIndex - 1 : guideIndex + 1;
-    return newGuideIndex > 0 && newGuideIndex < flatContents.length - 1 ? (
-        <Button>
-          <Link to={flatContents[newGuideIndex].frontmatter.path}>{direction} Guide</Link>
-        </Button>
-      ) : null
-    
-  }
 
   return (
     <Layout>
@@ -39,9 +25,7 @@ const Template = ({ data }) => {
             className="documentation-contents"
             dangerouslySetInnerHTML={{ __html: post.html }}
           />
-          {directionalButton('Previous')}
-          {directionalButton('Next')}
-          <a href={`${gitHubUrl}${post.frontmatter.path}.md`} target="_blank" rel="noopener noreferrer">Contribute to this doc</a>
+          <DocFooter contents={contentsArray} currentGuide={post.frontmatter.path} />
         </div>
         <div style={{ maxWidth: '30%'}}>
           <RightMenu tableOfContents={contentsArray} />
@@ -73,7 +57,6 @@ export const pageQuery = graphql`
             section
             title
             path
-            priority
           }
         }
       }
